gpaste-file-backend: Implement routine to save an item in the end of file
Every time user copies something, this routine is called to save the last
added item to the history, overwritting the whole file, which can take a
few seconds if history is big.

Add g_paste_file_backend_write_history_last_item() to be called in those
cases. It just adds the new item in the end of the history file.

Signed-off-by: Leonardo Bras <leobras@redhat.com>